Move col class to static function
In order to work properly with ScrollToShow, the col class has to be on the same div as margin-center. ScrollToShow needs to be the parent of the div.col.margin-center because otherwise the scroll won't take the margin into account and the props.mainContent div will show later than props.children. This keeps both mainContent's and children's show-threshold the same