Use a random value for the clipPath.id
If multiple instances using the same ID are created, only the first one will be used throughout the app. If that first instance is created in a 'display: none;' div, then no other svg's using that pattern will be able to display the polygon. Thus, use a unique ID for all polygons