Extract same-height context logic to separate component and update parents accordingly
In order to make the same-height logic optional within ImageCard, there would either be lots of conditional logic to check if a Provider is used or not, or an extra prop 'isUsingSameHeightProvider'. Instead of adding all that additional logic, extract it to an HOC to handle. Also, add extra 'onLoad' and 'imageStyle' props to ImageCard so the same-height fields can be passed to the correct children