Replace setting div width manually with auto CSS width fitting
The original problem this aimed to solve was the image having a smaller width than the containing div, in which case the hover-content div would have needed its width to be set to match the image width. Previously, the calculation of width was based off when the image loaded which didn't account for the times when the window was resized. This accomplishes both of those by forcing the entire containing div to fit the content width regardless of when that width changes (on load and on window resize).