Accept an optional imageRef from the parent
Previously, in the case where the parent component needs to know about the inner image element, the inner ImageCard's imageRef was passed to props.onLoad(). To avoid passing child state up to the parent in this manner, now the ImageCard can take an optional imageRef as a prop so the parent can do all the handling it needs.