Bug fix: use imageRef instead of parentDivRef for height calculation
In the event that the first image in SameHeightImageCard.Provider is larger than the following ones, the parentDivRef's height in the latter components would be larger than the rendered image's height, so the overall context height would not be updated to reflect the smallest image's height. Using a ref to the image instead of the parent div fixes this.