Replace ImageCard's manual, hard-coded margins with automatic flex space-evenly
In some particular screen sizes, the manual margins caused the text to either be too close to the ImageCard's text wrapper's bottom or to run off it completely. Since these manual margins force the title/description children to be spaced exactly with those margins regardless of parent div size, replacing it with flex-box's `align-content: space-evenly` allows the title/description children to be spaced evenly within the text wrapper div dynamically. Also, in the process of removing the manual margins, add a generic `margin-top` for the title and center both title and description horizontally