Remove hover-expand animation in mobile (not tablet) view
In mobile (not tablet) views, the hover interaction is replaced by a screen tap. However, with the smaller screen size, expanding the Projects' InfoCard entries upon tap makes them too wide, and is thus not a pleasant viewing experience, so remove it when on mobile. With tablets, this isn't an issue, so keep the expansion upon tap as-is