Improve performance by removing unnecessary scroll event listeners after they're completed
ScrollToShow is used often throughout the app, so many scroll event listeners are being added to the window. However, once all the children passed to ScrollToShow are shown, there will no longer be any work done on them, so there is no reason for the scroll listener to continue listening. Thus, remove it once it finishes its intended purpose