Add ability for <Shape/> to render centered text that changes font size dynamically
Previously, any text desired to be rendered on top of <Shape/> would require using `position: absolute;` and manually choosing font size, margin spacing, etc. However, this doesn't handle dynamic text sizes based on the size of the rendered SVG, which causes problems with some screen sizes. Putting the text inside the SVG allows the text to change font size based on the size of the SVG, forcing the text to always be the correct size.