Move atom_spinner.svg content to React component
Because this spinner is meant to be the app-wide page-loading spinner, it needs to resolve immediately before other images resolve. To do this, instead of importing the .svg file, put the <svg/> straight into a React component and import that instead. This way, the spinner will load/display immediately without needing to resolve image/file imports.