Add logic to display the global spinner when images are still loading
The end goal is to display the app-wide spinner while images are still loading. To do so, instead of adding onLoad handlers from the parent to every child <Image/>, simply read the app context from the Image component itself and update the number of requested/loaded images respectively. Similarly, add an optional 'updateAppContext' prop to disable this functionality in the event we don't care if a particular image loads before displaying the page. Finally, when determining if the spinner should show, avoid flashing the app (instead of spinner) upon initial JavaScript load but before the images have been requested by adding an edge-case check for `imagesHaveBeenRequested`