Hide app content while spinner is still showing
Originally, scrolling was attempted to be captured using `overscroll-behavior: contain;` on the spinner's rendered div, but it was not preventing the user from scrolling because there was nowhere to scroll in the spinner (it takes up full screen height/width, not more than that to afford scrolling). To get around this, make the spinner have a fixed position and scroll to the top of the page once it is supposed to be hidden.