Prevent scrolling when app-wide spinner is active
Previously, the main app content was only hidden when the spinner was active, but this still allowed the user to scroll. While the app would scroll back to the top after the spinner was removed, it would activate any animations that they happened to scroll past, which ruined the experience of those animations. To fix this, add an optional 'preventScrolling' to <SpinnerAtom/> that automatically scrolls to the top of the page if the user scrolls only while the spinner is active.