Improve accuracy of dynamic font shrinking hook
Previously, the height/width ratios were applied to the font size (e.g. if height is 2 times too big, then halve the font-size). However, font-size doesn't work this way and cannot be compared to an exact height/width px conversion. So, replace it with reducing the font-size by 1 px at a time, and if it's still too big, then keep reducing it (done by adding 'fontSizePx' to the useEffect dependency array)