Allow `Shape.props.text` to be either a node or a function so the font-resize ref can be applied to specific child
Previously, the font-resize ref was applied directly to the foreignObject's div. However, this doesn't help reduce font-size when the `props.text` node has `position: absolute` styling because the foreignObject's div's height doesn't change in this situation (absolute elements are outside the standard HTML flow). Thus, allow the parent to pass a function so they can place the ref on the correct child if needed.