Add calculations for getting the foreignObject's `boundingClientRect()` relative to window
Since SVGs have their own viewport, the standard `foreignObject.getBoundingClientRect()` call will return dimensions relative to the SVG. However, to use the bounding rect for any functions outside the SVG, this will give inaccurate results because any functions outside the SVG will expect the bounding rect to be relative to the window. Thus, calculate the bounding rect relative to the window by hand and then pass it to the `Shape.props.htmlChildren` function so the parent can use the correct foreignObject's rect. Also, add a comment in the PropTypes declaration to show this usage.