Bug fix: Update Shape's getBoundingClientRect() values on scroll
Previously, the SVG's bounding rect was not updated after the page loaded (which is also why it took the window.pageYOffset into account inside the 'useHover()' hook). This didn't cause an issue on desktops, but it skewed the position calculation for mobile devices. Fix this by adding a scroll listener to update the SVG's bounding client rect on scroll