Move favicon to root
Many browsers automatically search for the favicon at the root of the website. While it functions totally fine with the favicon nested in the assets folder, it would be better to conform to web-dev standards for compatibility than make something particularly custom and risk a browser not supporting it. Also, add the 'shortcut' field to the index.html's favicon link.rel for showing the image upon installation on mobile devices