Make each section have the same background color as title
In order to make the section flow a bit cleaner and to remove the stark section title-content separation, make the title and content portions of each section the same color, but alternate the colors to show the differences between sections. Also, in the process of doing this, remove the extra spacing between each section's title and content since the title no longer is as starkly separated from the content (note that h1 elements have the same margin-bottom as .mb-2. Finally, add an extra SCSS color for somewhere between light and lighter, and apply that in-between color to the MockRequests background (which involves renaming 'lighter' to 'lightest' and updating the body bg-color accordingly)