Add fetch error handler to ContactModal
Previously, the form error only displayed if fetch() was successful but the response was not (e.g. HTTP 400 error). However, this didn't handle the case where fetch() failed to send (e.g. the case where the user is offline), which resulted in the button spinning forever. Fix that by adding a try/catch around fetch().