Add specific form-wide error message for network failures
If the user is not connected to the internet, then fetch() will throw an error. In this case, show an error message that's specific to network failures instead of the generic one. Otherwise, if the Formspree submission failed, then show the original generic message to entice another submission attempt from the user.