Extract CSS class .duration-XX parsing to util function
This makes the CSS class parsing a bit more official by putting the class-parsing logic in one single location, so if the logic changes in the future, only one code block's logic needs to be updated. Also, allows for documentation to be added to the function for clarity of use.