Update animation duration time for .hover-expand after InfoCards slide in
After fixing the 'lazy-loading causing bootstrap CSS to overwrite custom CSS' issue from the previous commit, it became apparent that two animations were attached to the InfoCards (.slide-in-left and .hover-expand) but the animation/transition durations for each of those need to be different. Likewise, nested div.col-sm-4>div.hover-expand had odd effects where the box-shadow would be placed under the div.col-sm-4, causing the InfoCards rendered earlier to have shadows placed under the next InfoCard, so the CSS classes could not be separated into different divs. To fix both these issues, place the .duration-XX class in a state var and update it in ScrollToShow.props.onAllChildrenShown