Add animation for revealing easter egg
Nest shapes that render text inside a FlipCard component so that the entire shape flips when revealing the easter egg text. Stagger the flipping of shape entries using useTimedArrayToggle, and allow it to go forwards and backwards to see the original and easter egg texts. Finally, remove superfluous pageText variable extraction since that defeats the whole purpose of a pageText config object.