Add border and spacing between Production ImageCards
Previously, the ImageCards were displayed border-to-border, which gave a relatively clean experience. However, it wasn't clear to all users that the images would change on hover. Adding the border around the images reveals this to those users, and if a border is displayed, adding a bit of spacing around the border helps make it more appealing