Add new component for drawing SVG text on the screen along with associated CSS animations
Mixing `stroke-dasharray` and `stroke-dashoffset` in animations gives the net effect of drawing the outline of an SVG element on the screen. As explained [here](https://css-tricks.com/svg-line-animation-works/) this works by:

- Make the path outline/border a dashed line instead of solid line with `stroke-dasharray`.
- Make the `stroke-dasharray` the whole length of the path so that you only have one single dash.
- Use the `stroke-dashoffset` to set where the dash begins.
- Animate the `stroke-dashoffset` to go from length(`stroke-dasharray`) to 0 (or vice versa) to give the appearance of the single, long dash outlining the path.
- Mixing `stroke-dasharray` with an animated `stroke-dashoffset` gives the appearance of drawing the specified path
- Use `animation-fill-mode: forwards;` to make the final animation frame remain after animation is complete.

Since the `stroke-dasharray` length varies based on the element being drawn and because this is only a `text` element, we can base the length calculation off the font size. It comes out to about 75px per 1em font size. So, in order to make a re-usable component that calculates the `stroke-dasharray` length correctly regardless of font size, set the length value in a CSS variable and then update that variable in the `style` prop. This automatically updates both the length and the `@keyframes` entry for where to begin/end. Further, since that CSS variable is local to `.draw-text-and-fill`, one SvgDrawingText's font size won't affect another's.