Improve SpinnerAtom.props.preventScrolling behavior with direct styles
The original scroll event listener is only called on scroll event, which means a scroll must take place before it's fired. This means that it's technically possible for the user to scroll down the page a bit, even though after scrolling they would be returned to the top of the page. This has the effect of being able to ruin the rendered sections' animations when they initially appear. Thus, replace that buggy event handler with direct document.body styling to prevent them from being able to scroll at all