Update SameHeightImageCard to resize images on window resize
By changing the context state to use an array instead of the smallest image height, the index of the SameHeightImageCard can be used to update the context height value that corresponds to it. Doing so allows the context's heights to update individually without colliding with other component's state modifications and allows the setContextState function to be called at any point in time (necessary because useState() is both asynchronous and doesn't offer the second callback field parameter like class components do). Finally, remove the forced window.location.reload() logic since it's no longer needed