Fit MockRequests image height in ImageCard to show whole icon
Previously, the MockRequests image was longer than it should be to account for the image not centering appropriately in the ImageCard. This could easily be fixed by cropping the image to only include the icon without the extra space at the bottom and styling the <img> with 'height: 100%'. Relatedly, don't force the width to be 100% so the height dictates the image's display