Bug fix: Fix Safari cutting off a portion of the SplashSection text
Safari is the new IE. As such, it was incorrectly cutting off the left-most portion of the 'H' in 'Hey there' in the SvgDrawingText component. Nesting the <svg> (with a new width/height of 100%) in a <div> (which holds the actual width/height restrictions of 80%/75% respectively) corresponds with the `viewBox="0 0 100 100"` to not cut off the left-most portion of 'H'.