Add ability to allow client and service worker to communicate
The BroadcastChannel is a class that allows for posting and receiving messages globally across the whole app. Adding functions in ServiceWorker.js, registerServiceWorker.js, and Hooks.js allows each of them to access a BroadcastChannel that is used in the rest of the app. Similarly, configure the global BroadcastChannel name in webpack.config.js: `process.env.BROADCAST_CHANNEL` for client code, string-injected for ServiceWorker.