Improve App component's performance via memoizing; split components for context
Like SplashSection(Provider), since the App component depends on context, it needs to be split into two components (a memoized component and one that reads the context). To do so, split App.js into App.js and AppProvider.js where App is memoized and AppProvider reads the context. Likewise, move IncompatibleBrowserFallback to App/ since App is the only component that uses it.