Add all new files from the build to git in CD deploy step
Some files are added and some removed. As such, they should be added to the list of deployed files. Do so by iterating through the build output dir (`dist/`) and adding those files with the same name that were copied to root (since the build output dir is ignored in git).