Shrink Projects' entries on xs screen sizes
This one was a bit tricky due to how the images, text, etc. are all different sizes. A refactor here is definitely needed to make this style-logic simpler and flow automatically.

The problem was that if the title or text description of a Project entry were too small or two big, it resulted in more/fewer lines of text being present under the image, which pushed the image up or down, misaligning it with the other entries' images.

Until we make a full refactor, the temporary changes made to fix this include:

* Replace 'Polypeptides' with 'Proteins' for a shorter title length -- since 'Polypeptides' can't word-wrap as often as 'Atoms of Confusion', the easy solution is to just shorten word char length so it wraps at the same time the other wraps.
* Shrink `font-size` of 'Tardigrade Stratification' -- Fixing the 'Polypeptides' issue above only revealed a similar problem with the Tardigrade project. Since neither of these two words can be shortened, shrinking the `font-size` was the only solution. Until a cleaner/dynamic code change can be made, we hard-code the largest font size available without causing a word-wrap onto two lines in medium screen sizes.

Aside from those bugs, the originally-intended changes were:

* Shrink the width of Project entries by removing the `.w-100` parent element and adding `.col-11` to each `InfoCard` entry's parent div to make its width 11 cols instead of 12 but only on < sm screen sizes.
* Update the parent element of the one we just removed with `.flex-center` to ensure the newly modified, smaller-width `InfoCard` entries would be centered (wasn't a problem before since they took up the full container's width).