Change contact image screen-entry animation on mobile for slow devices
Sometimes, the images would lag when flipping, causing them to lag slightly upon first visual appearance (showing just a small line) and then suddenly jump to a full-width image. This was debugged extensively, but the cause varied; the only conclusive cause I could find was that it depended on if the page had loaded before (and was therefore cached), if the user scrolled down really fast (causing the CPU to be split between animations they already scrolled past and the current, Personal Contact image-flipping animations), and how fast their device was. Thus, in order to avoid this issue occurring for users, the simplest fix is to just replace the flip animation with a different one (in this case, I chose slide since it matched most of the rest of the site).