Bug fix: Make `<SameHeightImageCard/>` grow if screen space allows larger images to be rendered
Previously, the context state logic would only ever shrink images. The problem with that is if the images were shrunk upon first resize (e.g. if window width is decreased), then the images' new min/max height would be constricted to that shrunk size. If the window width were increased afterwards such that theoretically the image heights could also increase, they'd either stay the same or *decrease further* since the `<ImageCard/>` height was set to that previous, smaller height. This meant the images would rarely be the correct size that they could be on a given screen size, and caused them to repetitively shrink upon window width grow *and* shrink.

Fix this by creating a new "ghost" `<ImageCard/>` with most of the same styling as the "real" `<ImageCard/>` except with `height: auto` to see what the height would have actually been if we had just rendered the image naturally. This way, we always get an accurate reading of what the max/min heights of the images could be and can thus make an informed, correct decision on what the final heights should be.