Add prop to `<Modal/>` for rendering it via Portal
Sometimes, the parent DOM elements' styles can affect child styles even if the child is displayed in such a way that it shouldn't be affected (e.g. modals with `position: fixed|absolute`). This prop allows us to render a modal outside of the app's root DOM element so that we ensure no negative side effects from parents' styling affects the modal.