WIP: Remove `import-assertions` ESLint rule since it's conflicting with `import-attributes`
I think this is only an issue because some combination of Babel, TypeScript, and ESLint (itself or the plugins we're using) are outdated and conflicting with each other. In other apps, esp `react-app-boilerplate`, this exact ESLint config is working perfectly. Since I don't really have the time right now to do a deep dive, I'll save this as a TODO and revisit later.